requirements.txt
youtube-dl
selenium
requests
cryptography
Pillow (Python Image Libary)
------------------------------------
ffmpeg is needed in command line.